requirements.txt
aniso8601==7.0.0
arrow==1.3.0
asgiref==3.7.2
Babel==2.13.1
certifi==2023.7.22
channels==4.0.0
charset-normalizer==3.3.2
Django==5.0b1
django-admin-tools==0.9.3
django-currentuser==0.6.1
django-filter==23.3
django-money==3.4.0
django-phonenumber-field==7.2.0
django-rest-framework==0.1.0
django-urls==1.1.3
djangorestframework==3.14.0
drf-yasg==1.21.7
form==0.0.1
graphene==2.1.9
graphene-django==2.16.0
graphene-djmoney==0.2.3
graphql-core==2.3.2
graphql-relay==2.0.1
idna==3.4
inflection==0.5.1
numpy==1.26.2
opencv-python==4.8.1.78
packaging==23.2
phonenumbers==8.13.26
Pillow==10.1.0
promise==2.3
psycopg2-binary==2.9.9
py-moneyed==3.0
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
Rx==1.6.3
singledispatch==4.1.0
six==1.16.0
sqlparse==0.4.4
text-unidecode==1.3
types-python-dateutil==2.8.19.14
typing_extensions==4.8.0
uritemplate==4.1.1
urllib3==2.0.7
Yag==1.0.0
